+++ src/onegov.gazette/onegov/gazette/tests/test_layout.py
@@ -130,10 +130,13 @@ def test_layout_format(session, principal):
         layout.format_issue('')
 
     assert layout.format_issue(Issue()) == 'No. , '
-    assert layout.format_issue(Issue(number=1, date=date(2017, 1, 2))) \
-        == 'No. 1, 02.01.2017'
+
     assert layout.format_issue(
-        Issue(number=1, date=date(2017, 1, 2)), date_format='date_with_weekday'
+        Issue(number=1, date=date(2017, 1, 2))
+    ) == 'No. 1, 02.01.2017'
+    assert layout.format_issue(
+        Issue(number=1, date=date(2017, 1, 2)),
+        date_format='date_with_weekday'
     ) == 'No. 1, Montag 02.01.2017'
     assert layout.format_issue(
         Issue(name='2017-1', number=1, date=date(2017, 1, 2)),